backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
